Skip functional test when pushing to main
The functional tests are a good gauge of whether we are replicating
`aqt list-qt` properly or not. However, they often time out, or fail
due to reasons completely outside this project's control (ie the
download.qt.io server is down or slow).

This change skips the functional tests when pushing to the main branch,
so that problems at the download.qt.io server do not affect deployments
for this project.