display module names properly
The schema for fully-qualified names for packages has changed with
Qt 6.7.0. In the past, it was sufficient to just split the name on `.`
and take the second-to-last substring:

* the name 'qt.qt6.670.qtshadertools.win64_llvm_mingw' becomes
  'qtshadertools'

That doesn't work for new packages like this:

* 'qt.qt6.680.addons.qt3d.debug_information.win64_llvm_mingw' becomes
  'debug_information'
* 'qt.qt6.670.qtshadertools.debug_information.win64_llvm_mingw' becomes
  'debug_information'

It is now more effective to use a regular expression to pick out the
module name.